\name{new1-package}
\alias{new1-package}
\alias{new1}
\docType{package}
\title{
\packageTitle{new1}
}
\description{
\packageDescription{new1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{new1}
\packageIndices{new1}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{new1}

Maintainer: \packageMaintainer{new1}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
